+++ jib-core/src/main/java/com/google/cloud/tools/jib/registry/RegistryClient.java
@@ -23,7 +23,7 @@ import com.google.cloud.tools.jib.blob.Blob;
 import com.google.cloud.tools.jib.blob.BlobDescriptor;
 import com.google.cloud.tools.jib.blob.Blobs;
 import com.google.cloud.tools.jib.builder.TimerEventDispatcher;
-import com.google.cloud.tools.jib.event.EventDispatcher;
+import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.cloud.tools.jib.global.JibSystemProperties;
 import com.google.cloud.tools.jib.http.Authorization;
 import com.google.cloud.tools.jib.image.json.BuildableManifestTemplate;
@@ -44,7 +44,7 @@ public class RegistryClient {
   /** Factory for creating {@link RegistryClient}s. */
   public static class Factory {
 
-    private final EventDispatcher eventDispatcher;
+    private final EventHandlers eventHandlers;
     private final RegistryEndpointRequestProperties registryEndpointRequestProperties;
 
     private boolean allowInsecureRegistries = false;
@@ -52,9 +52,9 @@ public class RegistryClient {
     @Nullable private Authorization authorization;
 
     private Factory(
-        EventDispatcher eventDispatcher,
+        EventHandlers eventHandlers,
         RegistryEndpointRequestProperties registryEndpointRequestProperties) {
-      this.eventDispatcher = eventDispatcher;
+      this.eventHandlers = eventHandlers;
       this.registryEndpointRequestProperties = registryEndpointRequestProperties;
     }
 
@@ -99,7 +99,7 @@ public class RegistryClient {
      */
     public RegistryClient newRegistryClient() {
       return new RegistryClient(
-          eventDispatcher,
+          eventHandlers,
           authorization,
           registryEndpointRequestProperties,
           allowInsecureRegistries,
@@ -132,18 +132,16 @@ public class RegistryClient {
   /**
    * Creates a new {@link Factory} for building a {@link RegistryClient}.
    *
-   * @param eventDispatcher the event dispatcher used for dispatching log events
+   * @param eventHandlers the event handlers used for dispatching log events
    * @param serverUrl the server URL for the registry (for example, {@code gcr.io})
    * @param imageName the image/repository name (also known as, namespace)
    * @return the new {@link Factory}
    */
-  public static Factory factory(
-      EventDispatcher eventDispatcher, String serverUrl, String imageName) {
-    return new Factory(
-        eventDispatcher, new RegistryEndpointRequestProperties(serverUrl, imageName));
+  public static Factory factory(EventHandlers eventHandlers, String serverUrl, String imageName) {
+    return new Factory(eventHandlers, new RegistryEndpointRequestProperties(serverUrl, imageName));
   }
 
-  private final EventDispatcher eventDispatcher;
+  private final EventHandlers eventHandlers;
   @Nullable private final Authorization authorization;
   private final RegistryEndpointRequestProperties registryEndpointRequestProperties;
   private final boolean allowInsecureRegistries;
@@ -152,18 +150,18 @@ public class RegistryClient {
   /**
    * Instantiate with {@link #factory}.
    *
-   * @param eventDispatcher the event dispatcher used for dispatching log events
+   * @param eventHandlers the event handlers used for dispatching log events
    * @param authorization the {@link Authorization} to access the registry/repository
    * @param registryEndpointRequestProperties properties of registry endpoint requests
    * @param allowInsecureRegistries if {@code true}, insecure connections will be allowed
    */
   private RegistryClient(
-      EventDispatcher eventDispatcher,
+      EventHandlers eventHandlers,
       @Nullable Authorization authorization,
       RegistryEndpointRequestProperties registryEndpointRequestProperties,
       boolean allowInsecureRegistries,
       String userAgent) {
-    this.eventDispatcher = eventDispatcher;
+    this.eventHandlers = eventHandlers;
     this.authorization = authorization;
     this.registryEndpointRequestProperties = registryEndpointRequestProperties;
     this.allowInsecureRegistries = allowInsecureRegistries;
@@ -224,7 +222,7 @@ public class RegistryClient {
     return Verify.verifyNotNull(
         callRegistryEndpoint(
             new ManifestPusher(
-                registryEndpointRequestProperties, manifestTemplate, imageTag, eventDispatcher)));
+                registryEndpointRequestProperties, manifestTemplate, imageTag, eventHandlers)));
   }
 
   /**
@@ -296,7 +294,7 @@ public class RegistryClient {
         new BlobPusher(registryEndpointRequestProperties, blobDigest, blob, sourceRepository);
 
     try (TimerEventDispatcher timerEventDispatcher =
-        new TimerEventDispatcher(eventDispatcher, "pushBlob")) {
+        new TimerEventDispatcher(eventHandlers, "pushBlob")) {
       try (TimerEventDispatcher timerEventDispatcher2 =
           timerEventDispatcher.subTimer("pushBlob POST " + blobDigest)) {
 
@@ -347,7 +345,7 @@ public class RegistryClient {
   private <T> T callRegistryEndpoint(RegistryEndpointProvider<T> registryEndpointProvider)
       throws IOException, RegistryException {
     return new RegistryEndpointCaller<>(
-            eventDispatcher,
+            eventHandlers,
             userAgent,
             getApiRouteBase(),
             registryEndpointProvider,