+++ jib-core/src/test/java/com/google/cloud/tools/jib/builder/ProgressEventDispatcherTest.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.tools.jib.builder;
 
-import com.google.cloud.tools.jib.event.EventDispatcher;
+import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.cloud.tools.jib.event.events.ProgressEvent;
 import java.util.List;
 import org.junit.Assert;
@@ -31,12 +31,12 @@ import org.mockito.junit.MockitoJUnitRunner;
 @RunWith(MockitoJUnitRunner.class)
 public class ProgressEventDispatcherTest {
 
-  @Mock private EventDispatcher mockEventDispatcher;
+  @Mock private EventHandlers mockEventHandlers;
 
   @Test
   public void testDispatch() {
     try (ProgressEventDispatcher progressEventDispatcher =
-            ProgressEventDispatcher.newRoot(mockEventDispatcher, "ignored", 10);
+            ProgressEventDispatcher.newRoot(mockEventHandlers, "ignored", 10);
         ProgressEventDispatcher ignored =
             progressEventDispatcher.newChildProducer().create(BuildStepType.ALL, "ignored", 20)) {
       // empty
@@ -44,7 +44,7 @@ public class ProgressEventDispatcherTest {
 
     ArgumentCaptor<ProgressEvent> progressEventArgumentCaptor =
         ArgumentCaptor.forClass(ProgressEvent.class);
-    Mockito.verify(mockEventDispatcher, Mockito.times(4))
+    Mockito.verify(mockEventHandlers, Mockito.times(4))
         .dispatch(progressEventArgumentCaptor.capture());
     List<ProgressEvent> progressEvents = progressEventArgumentCaptor.getAllValues();
 
@@ -60,14 +60,14 @@ public class ProgressEventDispatcherTest {
   @Test
   public void testDispatch_safeWithtooMuchProgress() {
     try (ProgressEventDispatcher progressEventDispatcher =
-        ProgressEventDispatcher.newRoot(mockEventDispatcher, "allocation description", 10)) {
+        ProgressEventDispatcher.newRoot(mockEventHandlers, "allocation description", 10)) {
       progressEventDispatcher.dispatchProgress(6);
       progressEventDispatcher.dispatchProgress(8);
       progressEventDispatcher.dispatchProgress(1);
     }
 
     ArgumentCaptor<ProgressEvent> eventsCaptor = ArgumentCaptor.forClass(ProgressEvent.class);
-    Mockito.verify(mockEventDispatcher, Mockito.times(4)).dispatch(eventsCaptor.capture());
+    Mockito.verify(mockEventHandlers, Mockito.times(4)).dispatch(eventsCaptor.capture());
     List<ProgressEvent> progressEvents = eventsCaptor.getAllValues();
 
     Assert.assertSame(progressEvents.get(0).getAllocation(), progressEvents.get(1).getAllocation());
@@ -85,7 +85,7 @@ public class ProgressEventDispatcherTest {
   @Test
   public void testDispatch_safeWithTooManyChildren() {
     try (ProgressEventDispatcher progressEventDispatcher =
-            ProgressEventDispatcher.newRoot(mockEventDispatcher, "allocation description", 1);
+            ProgressEventDispatcher.newRoot(mockEventHandlers, "allocation description", 1);
         ProgressEventDispatcher ignored1 =
             progressEventDispatcher.newChildProducer().create(BuildStepType.ALL, "ignored", 5);
         ProgressEventDispatcher ignored2 =
@@ -94,7 +94,7 @@ public class ProgressEventDispatcherTest {
     }
 
     ArgumentCaptor<ProgressEvent> eventsCaptor = ArgumentCaptor.forClass(ProgressEvent.class);
-    Mockito.verify(mockEventDispatcher, Mockito.times(5)).dispatch(eventsCaptor.capture());
+    Mockito.verify(mockEventHandlers, Mockito.times(5)).dispatch(eventsCaptor.capture());
     List<ProgressEvent> progressEvents = eventsCaptor.getAllValues();
 
     Assert.assertEquals(1, progressEvents.get(0).getAllocation().getAllocationUnits());