+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/GradleProjectProperties.java
@@ -100,7 +100,7 @@ class GradleProjectProperties implements ProjectProperties {
     }
     ConsoleLogger consoleLogger = consoleLoggerBuilder.build();
 
-    return new EventHandlers()
+    return EventHandlers.builder()
         .add(
             JibEventType.LOGGING,
             logEvent -> consoleLogger.log(logEvent.getLevel(), logEvent.getMessage()))
@@ -116,7 +116,8 @@ class GradleProjectProperties implements ProjectProperties {
                           update.getProgress(), update.getUnfinishedLeafTasks());
                   footer.add("");
                   consoleLogger.setFooter(footer);
-                }));
+                }))
+        .build();
   }
 
   private static boolean isProgressFooterEnabled(Project project) {