+++ jib-plugins-common/src/test/java/com/google/cloud/tools/jib/plugins/common/ConfigurationPropertyValidatorTest.java
@@ -19,7 +19,7 @@ package com.google.cloud.tools.jib.plugins.common;
 import com.google.cloud.tools.jib.api.ImageReference;
 import com.google.cloud.tools.jib.api.InvalidImageReferenceException;
 import com.google.cloud.tools.jib.configuration.credentials.Credential;
-import com.google.cloud.tools.jib.event.EventDispatcher;
+import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.cloud.tools.jib.event.events.LogEvent;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -35,7 +35,7 @@ import org.mockito.junit.MockitoJUnitRunner;
 @RunWith(MockitoJUnitRunner.class)
 public class ConfigurationPropertyValidatorTest {
 
-  @Mock private EventDispatcher mockEventDispatcher;
+  @Mock private EventHandlers mockEventHandlers;
   @Mock private AuthProperty mockAuth;
   @Mock private ImageReference mockImageReference;
   @Mock private RawConfiguration mockConfiguration;
@@ -55,7 +55,7 @@ public class ConfigurationPropertyValidatorTest {
     Credential expected = Credential.from("abcde", "12345");
     Optional<Credential> actual =
         ConfigurationPropertyValidator.getImageCredential(
-            mockEventDispatcher,
+            mockEventHandlers,
             "jib.test.auth.user",
             "jib.test.auth.pass",
             mockAuth,
@@ -69,21 +69,21 @@ public class ConfigurationPropertyValidatorTest {
     expected = Credential.from("vwxyz", "98765");
     actual =
         ConfigurationPropertyValidator.getImageCredential(
-            mockEventDispatcher,
+            mockEventHandlers,
             "jib.test.auth.user",
             "jib.test.auth.pass",
             mockAuth,
             mockConfiguration);
     Assert.assertTrue(actual.isPresent());
     Assert.assertEquals(expected.toString(), actual.get().toString());
-    Mockito.verify(mockEventDispatcher, Mockito.never()).dispatch(LogEvent.warn(Mockito.any()));
+    Mockito.verify(mockEventHandlers, Mockito.never()).dispatch(LogEvent.warn(Mockito.any()));
 
     // Auth completely missing
     Mockito.when(mockAuth.getUsername()).thenReturn(null);
     Mockito.when(mockAuth.getPassword()).thenReturn(null);
     actual =
         ConfigurationPropertyValidator.getImageCredential(
-            mockEventDispatcher,
+            mockEventHandlers,
             "jib.test.auth.user",
             "jib.test.auth.pass",
             mockAuth,
@@ -95,13 +95,13 @@ public class ConfigurationPropertyValidatorTest {
     Mockito.when(mockAuth.getPassword()).thenReturn(null);
     actual =
         ConfigurationPropertyValidator.getImageCredential(
-            mockEventDispatcher,
+            mockEventHandlers,
             "jib.test.auth.user",
             "jib.test.auth.pass",
             mockAuth,
             mockConfiguration);
     Assert.assertFalse(actual.isPresent());
-    Mockito.verify(mockEventDispatcher)
+    Mockito.verify(mockEventHandlers)
         .dispatch(
             LogEvent.warn("pass is missing from build configuration; ignoring auth section."));
 
@@ -110,13 +110,13 @@ public class ConfigurationPropertyValidatorTest {
     Mockito.when(mockAuth.getPassword()).thenReturn("98765");
     actual =
         ConfigurationPropertyValidator.getImageCredential(
-            mockEventDispatcher,
+            mockEventHandlers,
             "jib.test.auth.user",
             "jib.test.auth.pass",
             mockAuth,
             mockConfiguration);
     Assert.assertFalse(actual.isPresent());
-    Mockito.verify(mockEventDispatcher)
+    Mockito.verify(mockEventHandlers)
         .dispatch(
             LogEvent.warn("user is missing from build configuration; ignoring auth section."));
   }
@@ -138,19 +138,18 @@ public class ConfigurationPropertyValidatorTest {
     // Target configured
     ImageReference result =
         ConfigurationPropertyValidator.getGeneratedTargetDockerTag(
-            "a/b:c", mockEventDispatcher, "project-name", "project-version", helpfulSuggestions);
+            "a/b:c", mockEventHandlers, "project-name", "project-version", helpfulSuggestions);
     Assert.assertEquals("a/b", result.getRepository());
     Assert.assertEquals("c", result.getTag());
-    Mockito.verify(mockEventDispatcher, Mockito.never())
-        .dispatch(LogEvent.lifecycle(Mockito.any()));
+    Mockito.verify(mockEventHandlers, Mockito.never()).dispatch(LogEvent.lifecycle(Mockito.any()));
 
     // Target not configured
     result =
         ConfigurationPropertyValidator.getGeneratedTargetDockerTag(
-            null, mockEventDispatcher, "project-name", "project-version", helpfulSuggestions);
+            null, mockEventHandlers, "project-name", "project-version", helpfulSuggestions);
     Assert.assertEquals("project-name", result.getRepository());
     Assert.assertEquals("project-version", result.getTag());
-    Mockito.verify(mockEventDispatcher)
+    Mockito.verify(mockEventHandlers)
         .dispatch(
             LogEvent.lifecycle(
                 "Tagging image with generated image reference project-name:project-version. If you'd "
@@ -160,7 +159,7 @@ public class ConfigurationPropertyValidatorTest {
     // Generated tag invalid
     try {
       ConfigurationPropertyValidator.getGeneratedTargetDockerTag(
-          null, mockEventDispatcher, "%#&///*@(", "%$#//&*@($", helpfulSuggestions);
+          null, mockEventHandlers, "%#&///*@(", "%$#//&*@($", helpfulSuggestions);
       Assert.fail();
     } catch (InvalidImageReferenceException ignored) {
     }