+++ jib-core/CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Multiple classes have been moved to the `com.google.cloud.tools.jib.api` package
+- Removed `EventDispatcher` and `DefaultEventDispatcher`; events are now dispatched directly from `EventHandlers`
+
 ### Fixed
 
 ## 0.9.2